backend/requirements.txt
django-cors-headers==3.13.0
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.25.0
Faker==15.3.4
Markdown==3.4.1
Pillow==9.3.0
pandas==1.5.2
requests==2.28.1
scipy==1.9.3

